Moved tests to DateTimeFunction.cs and made them applicable to all providers
